chore: move index validation to BaseEmbedder, add empty GTFS2VecEmbedder class
