+++ srai/embedders/count_embedder.py
@@ -8,8 +8,10 @@ from typing import List, Optional
 import geopandas as gpd
 import pandas as pd
 
+from srai.embedders import BaseEmbedder
 
-class CountEmbedder:
+
+class CountEmbedder(BaseEmbedder):
     """Simple Embedder that counts occurences of feature values."""
 
     def __init__(self, expected_output_features: Optional[List[str]] = None) -> None:
@@ -82,41 +84,6 @@ class CountEmbedder:
         region_embedding_df = regions_df.join(region_embeddings, how="left").fillna(0).astype(int)
 
         return region_embedding_df
-
-    def _validate_indexes(
-        self,
-        regions_gdf: gpd.GeoDataFrame,
-        features_gdf: gpd.GeoDataFrame,
-        joint_gdf: gpd.GeoDataFrame,
-    ) -> None:
-        if regions_gdf.index.name is None:
-            raise ValueError("regions_gdf must have a named index.")
-
-        if features_gdf.index.name is None:
-            raise ValueError("features_gdf must have a named index.")
-
-        if not isinstance(joint_gdf.index, pd.MultiIndex):
-            raise ValueError(
-                f"joint_gdf.index must be of type pandas.MultiIndex, not {type(joint_gdf.index)}"
-            )
-
-        print(joint_gdf.index.names)
-        if len(joint_gdf.index.names) != 2:
-            raise ValueError(
-                f"joint_gdf.index must have 2 levels, has {len(joint_gdf.index.names)}"
-            )
-
-        if regions_gdf.index.name != joint_gdf.index.names[0]:
-            raise ValueError(
-                f"Name of regions_gdf.index ({regions_gdf.index.name}) must be equal to the name of"
-                f" the 1st level of joint_gdf.index ({joint_gdf.index.names[0]})"
-            )
-
-        if features_gdf.index.name != joint_gdf.index.names[1]:
-            raise ValueError(
-                f"Name of features_gdf.index ({features_gdf.index.name}) must be equal to the name"
-                f" of the 2nd level of joint_gdf.index ({joint_gdf.index.names[1]})"
-            )
 
     def _filter_to_expected_features(self, region_embeddings: pd.DataFrame) -> pd.DataFrame:
         """
